+++ Sources/SwiftExtras/Extensions/Color+default.swift
@@ -9,7 +9,7 @@
 //  MIT License
 //
 
-#if !os(watchOS) && canImport(SwiftUI) && (canImport(AppKit) || canImport(UIKit)) && !canImport(Colors)
+#if !os(watchOS) && canImport(SwiftUI) && (canImport(AppKit) || canImport(UIKit))
 import SwiftUI
 
 #if canImport(AppKit)